Add files via upload
This project was developed as part of the "Object-Oriented Programming" course in the postgraduate program "Control and Computing Courses". The assignment involved creating an RPG, focusing on implementing key components using object-oriented programming principles.

The project showcases a small RPG game where players can experience various game mechanics such as character creation, movement, and interaction with the game environment. The architecture is modular and object-oriented, demonstrating the use of classes along with inheritance, encapsulation, and polymorphism.